Make wiktionary parser self contained.
In previous design the fetching of data was done by bridge and passed to
wiktionary parser, but this make bridge to hold unnecessary
information. So we moved the fetching of data from wiktionary to
wiktionary parser so there by making it self contained. Bridge will just
pass data and expects output in some format.

Signed-off-by: Vasudev Kamath <kamathvasudev@gmail.com>